requirements.txt
Django==2.2
dj-database-url==0.5.0
psycopg2==2.7.7
whitenoise[brotli]==4.1.2
django-storages==1.7.1
celery==4.3.0
redis==3.2.1
boto3==1.9.128
raven==6.10.0
django-referrer-policy==1.0
djangorestframework==3.9.2
